backend/requirements.txt
boto3
requests
bcrypt
pyjwt
transbank-sdk
